feat: add log file for backup history

+++ backup.sh
@@ -5,6 +5,7 @@ set -a
 source .env
 set +a
 
+backup_count=0
 
 # Check if jq is installed
 if ! command -v jq &> /dev/null; then
@@ -18,19 +19,36 @@ if [ -z "$DEST_DIR" ]; then
     exit 1
 fi
 
-#Check if the folder exists
+# Check if the folder exists
 mkdir -p "$DEST_DIR"
 
+# Check if folders is setup
+if [ ! -f ".folders.json" ]; then
+  echo "Error: .folders.json not found!"
+  exit 1
+fi
 
-#Initialise the timestamp of the archive
+
+# Initialise the timestamp of the archive
 TIMESTAMP=$(date "+%Y-%m-%d_%H-%M-%S")
 
 # Read the folders you want to sync, from .folders.json
 FOLDERS=$(jq -r '.folders[]' .folders.json)
 
+LOG_FILE="./backup.log"
+
+# Create log file if it doesn't exist
+if [ ! -f "$LOG_FILE" ]; then
+  touch "$LOG_FILE"
+fi
+
+log() {
+  echo "[$(date '+%Y-%m-%d %H:%M:%S')] $1" >> "$LOG_FILE"
+}
+
 CHECKSUM_FILE=./last-checksums.json
 
-#Check if the checksum file exists
+# Check if the checksum file exists
 if [ ! -f "$CHECKSUM_FILE" ] || ! jq -e . "$CHECKSUM_FILE" > /dev/null 2>&1; then
   echo "{}" > "$CHECKSUM_FILE"
 fi
@@ -64,13 +82,17 @@ perform_backup()
       -czf "$archive_name" \
       -C "$(dirname "$folder_expanded")" "$basename"
 
+
+  log "Backup created for $folder → $(realpath "$archive_name")"
+  backup_count=$((backup_count + 1))
+
   #echo "Compression finished, sending to Proton Drive..."
   rsync -av "$archive_name" "$DEST_DIR/"
 
   rm "$archive_name"
   #echo "Backup completed : $basename"
 }
-
+log "=== New backup session started ==="
 for FOLDER in $FOLDERS; do
   FOLDER_EXPANDED=$(eval echo "$FOLDER")
   if [ -d "$FOLDER_EXPANDED" ]; then
@@ -84,13 +106,22 @@ for FOLDER in $FOLDERS; do
         '. + {($path): $sum}' "$CHECKSUM_FILE" > tmp.json && mv tmp.json "$CHECKSUM_FILE"
     else
       echo "$FOLDER already uptodate → skipping backup."
+      log "$FOLDER already uptodate → skipping backup."
     fi
   else
+    log "Folder not found: $FOLDER"
     echo "Folder not found: $FOLDER"
   fi
 done
 
-echo "All backup completed"
+if [ "$backup_count" -gt 0 ]; then 
+  echo "All backups completed"
+  log "Backup finished – $backup_count archive(s) created"
+else
+  echo "Everything is already up to date"
+  log "=== No backup done, all folders up to date ==="
+fi
+
